fix plots
